Refactor rregister lifecycle helper function
