+++ tests/integration/controller/kyma/kyma_moduleinfo_test.go
@@ -48,7 +48,7 @@ var _ = Describe("Kyma module control", Ordered, func() {
 	kyma.Spec.Modules = append(
 		kyma.Spec.Modules, module)
 
-	RegisterDefaultLifecycleForKyma(kyma)
+	RegisterDefaultLifecycleForKyma(ctx, controlPlaneClient, kyma)
 
 	DescribeTable("Test Manifests",
 		func(givenCondition func() error, expectedBehavior func() error) {
