+++ tests/integration/controller/kyma/moduletemplate_test.go
@@ -105,7 +105,7 @@ var _ = Describe("ModuleTemplate.Spec.descriptor not contains RegistryCred label
 	kyma.Spec.Modules = append(
 		kyma.Spec.Modules, NewTestModule("test-module", v1beta2.DefaultChannel))
 
-	RegisterDefaultLifecycleForKymaWithoutTemplate(kyma)
+	RegisterDefaultLifecycleForKymaWithoutTemplate(ctx, controlPlaneClient, kyma)
 
 	It("expect Manifest.Spec.installs not contains credSecretSelector", func() {
 		DeployModuleTemplates(ctx, controlPlaneClient, kyma)
@@ -119,7 +119,7 @@ var _ = Describe("ModuleTemplate.Spec.descriptor contains RegistryCred label", O
 	module := NewTestModule("test-module", v1beta2.DefaultChannel)
 	kyma.Spec.Modules = append(kyma.Spec.Modules, module)
 
-	RegisterDefaultLifecycleForKymaWithoutTemplate(kyma)
+	RegisterDefaultLifecycleForKymaWithoutTemplate(ctx, controlPlaneClient, kyma)
 
 	It("expect Manifest.Spec.installs contains credSecretSelector", func() {
 		template := builder.NewModuleTemplateBuilder().
